Refactor sidebar and routing to replace 'bulk-create' with 'inputs' and 'outputs'. Update styles for new two-column layout in settings. Remove legacy bulk import functionality and adjust related tests and documentation accordingly.
