+++ src/lib/admin.ts
@@ -4,7 +4,7 @@ import type { RequestEvent } from '@sveltejs/kit';
 
 /**
  * Require the current user to be an admin. Call from page load or action.
- * Throws redirect to / if not admin.
+ * Throws redirect to /calendar if not admin.
  */
 export function requireAdmin(event: RequestEvent): string {
 	const userId = event.locals.userId;
@@ -14,7 +14,7 @@ export function requireAdmin(event: RequestEvent): string {
 	const db = getDatabase();
 	const row = db.prepare('SELECT tier FROM user WHERE id = ?').get(userId) as { tier: string } | undefined;
 	if (row?.tier !== 'admin') {
-		throw redirect(303, '/');
+		throw redirect(303, '/calendar');
 	}
 	return userId;
 }
