+++ src/lib/components/Sidebar.svelte
@@ -1,17 +1,15 @@
 <script lang="ts">
 	import { fade } from 'svelte/transition';
-	import { theme, toggleTheme } from '$lib/stores/theme.js';
 	import { sidebarOpen, closeSidebar } from '$lib/stores/sidebar.js';
 	import { page } from '$app/stores';
 
 	const baseNavItems = [
-		{ href: '/', label: 'Home', icon: 'home' },
 		{ href: '/calendar', label: 'Calendar', icon: 'calendar' },
 		{ href: '/posts', label: 'Posts', icon: 'doc' },
 		{ href: '/schedules', label: 'Schedules', icon: 'clock' },
-		{ href: '/reports', label: 'Reports', icon: 'chart' },
-		{ href: '/bulk-create', label: 'Import', icon: 'stack' },
-		{ href: '/settings', label: 'Settings', icon: 'gear' }
+		{ href: '/inputs', label: 'Inputs', icon: 'stack' },
+		{ href: '/outputs', label: 'Outputs', icon: 'send' },
+		{ href: '/reports', label: 'Reports', icon: 'chart' }
 	];
 	const userTier = $page.data.userTier as string | null | undefined;
 	const navItems = $derived(
@@ -20,11 +18,11 @@
 
 	function iconPath(icon: string) {
 		const paths: Record<string, string> = {
-			home: 'M3 12l2-2m0 0l7-7 7 7M5 10v10a1 1 0 001 1h3m10-11l2 2m-2-2v10a1 1 0 01-1 1h-3m-6 0a1 1 0 001-1v-4a1 1 0 011-1h2a1 1 0 011 1v4a1 1 0 001 1m-6 0h6',
 			calendar: 'M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z',
 			doc: 'M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z',
 			clock: 'M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z',
 			stack: 'M4 6h16M4 10h16M4 14h16M4 18h16',
+			send: 'M6 12L3.269 3.126A59.768 59.768 0 0121.485 12 59.77 59.77 0 013.27 20.876L5.999 12zm0 0h7.5',
 			chart: 'M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z',
 			gear: 'M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z M15 12a3 3 0 11-6 0 3 3 0 016 0z',
 			users: 'M12 4.354a4 4 0 110 5.292M15 21H3v-1a6 6 0 0112 0v1zm0 0h6v-1a6 6 0 00-9-5.197M13 7a4 4 0 11-8 0 4 4 0 018 0z'
@@ -142,7 +140,7 @@
 	<div class="flex h-full flex-col">
 		<!-- Header: logo + collapse -->
 		<div class="flex h-24 items-center justify-between border-b border-[var(--sidebar-border)] px-8 ">
-			<a href="/" class="flex items-center gap-2 font-semibold text-[var(--sidebar-text)] w-full">
+			<a href="/calendar" class="flex items-center gap-2 font-semibold text-[var(--sidebar-text)] w-full">
 				<img src="/logo_text.svg" alt="PostPlan" class="mx-auto h-10 w-auto max-w-[200px] object-contain" />
 			</a>
 			<button
@@ -225,7 +223,9 @@
 			<p class="mb-2 px-3 text-xs font-semibold uppercase tracking-wider text-[var(--sidebar-text-muted)]">Main menu</p>
 			<ul class="space-y-0.5">
 				{#each navItems as item}
-					{@const isActive = $page.url.pathname === item.href || (item.href !== '/' && $page.url.pathname.startsWith(item.href + '/'))}
+					{@const isActive =
+						$page.url.pathname === item.href ||
+						(item.href !== '/calendar' && $page.url.pathname.startsWith(item.href + '/'))}
 					<li>
 						<a
 							href={item.href}
@@ -243,8 +243,8 @@
 				{/each}
 			</ul>
 		</nav>
-		<div class="border-t border-[var(--sidebar-border)] p-3 space-y-2">
-			{#if $page.data.session}
+		{#if $page.data.session}
+			<div class="border-t border-[var(--sidebar-border)] p-3">
 				<div class="px-3">
 					<a
 						href="/account"
@@ -270,34 +270,7 @@
 						</p>
 					</a>
 				</div>
-				<form method="POST" action="/auth/login?/signout" class="px-0">
-					<input type="hidden" name="options.redirectTo" value="/auth/login" />
-					<button
-						type="submit"
-						class="pressable-row flex w-full min-h-[44px] items-center gap-3 rounded-r-lg px-3 py-2.5 text-sm font-medium text-[var(--sidebar-text-muted)] hover:bg-[var(--sidebar-hover)] hover:text-[var(--sidebar-text)]"
-					>
-						Sign out
-					</button>
-				</form>
-			{/if}
-			<button
-				type="button"
-				class="pressable-row flex w-full min-h-[44px] items-center gap-3 rounded-r-lg px-3 py-2.5 text-sm font-medium text-[var(--sidebar-text-muted)] hover:bg-[var(--sidebar-hover)] hover:text-[var(--sidebar-text)]"
-				onclick={toggleTheme}
-				aria-label="Toggle theme"
-			>
-				{#if $theme === 'dark'}
-					<svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-						<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z" />
-					</svg>
-					Light
-				{:else}
-					<svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-						<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z" />
-					</svg>
-					Dark
-				{/if}
-			</button>
-		</div>
+			</div>
+		{/if}
 	</div>
 </aside>
