+++ src/routes/api/stripe/portal/+server.ts
@@ -1,4 +1,5 @@
 import { getDatabase } from '$lib/db/index.js';
+import { getStripeSecrets } from '$lib/server/stripeEnv.js';
 import { env } from '$env/dynamic/private';
 import Stripe from 'stripe';
 import { redirect } from '@sveltejs/kit';
@@ -9,7 +10,7 @@ export const GET: RequestHandler = async ({ locals }) => {
 	if (!session?.user?.id) {
 		throw redirect(303, '/auth/login');
 	}
-	const secretKey = env.STRIPE_SECRET_KEY;
+	const { secretKey } = getStripeSecrets();
 	if (!secretKey) {
 		return new Response('Stripe is not configured.', { status: 503 });
 	}
