+++ tests/e2e/import.spec.ts
@@ -1,8 +1,8 @@
 import { test, expect } from '@playwright/test';
 
-test.describe('Import (bulk-create)', () => {
-	test('bulk-create page requires auth', async ({ page }) => {
-		await page.goto('/bulk-create');
+test.describe('Inputs', () => {
+	test('inputs page requires auth', async ({ page }) => {
+		await page.goto('/inputs');
 		await expect(page).toHaveURL(/\/welcome/);
 	});
 });
