+++ vite.config.ts
@@ -49,8 +49,8 @@ export default defineConfig(({ mode }) => {
 					'**/node_modules/**',
 					// Comment-only barrel; no executable statements.
 					'src/lib/index.ts',
-					// Very large importer module: covered by importer unit tests, bulk-create server slice, and E2E.
-					'src/routes/bulk-create/+page.server.ts'
+					// Very large importer module: covered by importer unit tests, inputs server slice, and E2E.
+					'src/routes/inputs/+page.server.ts'
 				]
 			}
 		}
